pyMQTT/requirements.txt
paho-mqtt>=2.1.0
certifi>=2024.2.2
pymongo>=4.7.1